pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "selenium>=4.31.0",
    "webdriver-manager>=4.0.2",
]
